Handle previous migrations
