+++ .secrets.baseline
@@ -212,7 +212,7 @@
         "filename": "migrations/versions/e4c7b0ab68d3_create_tables.py",
         "hashed_secret": "adb1fcd33b07abf9b6a064745759accea5cb341f",
         "is_verified": false,
-        "line_number": 17
+        "line_number": 21
       }
     ],
     "tests/conftest.py": [
@@ -314,5 +314,5 @@
       }
     ]
   },
-  "generated_at": "2022-06-07T21:49:00Z"
+  "generated_at": "2022-06-23T20:33:51Z"
 }
